Changed some calculations for max keys in index node

+++ BPlusTree.h
@@ -26,7 +26,7 @@ class CNode
 {
 public:
     int ORDER_V;
-    int maxnum ;
+    short int maxnum;
     CNode();
     virtual ~CNode();
